+++ examples/rectangular_array_write_read.rs
@@ -142,7 +142,7 @@ fn rectangular_array_write_read() -> Result<(), Box<dyn std::error::Error>> {
     println!("The middle 4x2 subset is:\n{data_4x2}\n");
 
     // Show the hierarchy
-    let node = Node::new(&*store, "/").unwrap();
+    let node = Node::open(&store, "/").unwrap();
     let tree = node.hierarchy_tree();
     println!("The Zarr hierarchy tree is:\n{tree}");