+++ examples/sharded_array_write_read.rs
@@ -149,7 +149,7 @@ fn sharded_array_write_read() -> Result<(), Box<dyn std::error::Error>> {
     }
 
     // Show the hierarchy
-    let node = Node::new(&*store, "/").unwrap();
+    let node = Node::open(&store, "/").unwrap();
     let tree = node.hierarchy_tree();
     println!("The Zarr hierarchy tree is:\n{}", tree);