+++ examples/zip_array_write_read.rs
@@ -199,7 +199,7 @@ fn zip_array_write_read() -> Result<(), Box<dyn std::error::Error>> {
     read_array_from_store(array)?;
 
     // Show the hierarchy
-    let node = Node::new(&*store, "/").unwrap();
+    let node = Node::open(&store, "/").unwrap();
     let tree = node.hierarchy_tree();
     println!("The Zarr hierarchy tree is:\n{}", tree);