+++ tests/hierarchy.rs
@@ -1,11 +1,15 @@
+use std::sync::Arc;
+
 use zarrs::{node::Node, storage::store::FilesystemStore};
 
 #[test]
 fn hierarchy_tree() {
-    let store = FilesystemStore::new("./tests/data/hierarchy.zarr")
-        .unwrap()
-        .sorted();
-    let node = Node::new(&store, "/").unwrap();
+    let store = Arc::new(
+        FilesystemStore::new("./tests/data/hierarchy.zarr")
+            .unwrap()
+            .sorted(),
+    );
+    let node = Node::open(&store, "/").unwrap();
     let tree = node.hierarchy_tree();
     println!("{:?}", tree);
     assert_eq!(